--- client/src/containers/grid/filters/index.tsx
@@ -1,22 +1,0 @@
-"use client";
-
-import { useGetGridMeta } from "@/lib/grid";
-
-import GridFiltersItem from "@/containers/grid/filters/item";
-
-export default function GridFilters() {
-  const { data: gridMetaData } = useGetGridMeta();
-
-  return (
-    <div>
-      <h2 className="text-lg font-medium">Filters</h2>
-
-      <div className="space-y-2">
-        {gridMetaData &&
-          gridMetaData.datasets.map((dataset) => (
-            <GridFiltersItem key={dataset.var_name} {...dataset} />
-          ))}
-      </div>
-    </div>
-  );
-}
